[charon] refactor manifest parsing code
This is to allow the resource indexer to import
`convertManifestJsonToAvailableDatasetList` without the side-effects of
running code on import and setting `global.availableDatasets`.

No code changes, just extracting 2 functions into a new file.
